PACKAGE: Remove license stubs from the top of every file.
These notices are unnecessary to enforce GPL, they're ugly, and
they're a pain in the ass to keep updated. They do not enforce
anything that isn't already spelled out in the COPYING file.

My understanding is that GPL does not require that these duplicate
notices are preserved (only the overriding one in the COPYING file).